Split `participants` into multiple files
