+++ server/services/participants/status.ts
@@ -1,0 +1,11 @@
+import { dbClient, collections } from '../../db';
+
+export const invalidateStatus = async ({ currentStatusId }) =>
+  dbClient.db[collections.PARTICIPANTS_STATUS].update(
+    {
+      id: currentStatusId,
+    },
+    {
+      current: false,
+    }
+  );
